+++ packages/framework/attributor/src/runtimeAttributorDataStoreChannel.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { bufferToString, TypedEventEmitter } from "@fluid-internal/client-utils";
+import { TypedEventEmitter } from "@fluid-internal/client-utils";
 import { AttachState, IDeltaManager } from "@fluidframework/container-definitions/internal";
 import {
 	FluidObject,
@@ -29,10 +29,7 @@ import {
 	VisibilityState,
 	type ISummaryTreeWithStats,
 	type ITelemetryContext,
-	type AttributionInfo,
-	type AttributionKey,
 } from "@fluidframework/runtime-definitions/internal";
-import { SummaryTreeBuilder } from "@fluidframework/runtime-utils/internal";
 import {
 	ITelemetryLoggerExt,
 	MonitoringContext,
@@ -40,23 +37,25 @@ import {
 	createChildMonitoringContext,
 } from "@fluidframework/telemetry-utils/internal";
 
-import { OpStreamAttributor, type IAttributor } from "./attributor.js";
-import { opBlobName, type IRuntimeAttributor } from "./attributorContracts.js";
-import { AttributorSerializer, chain, deltaEncoder, type Encoder } from "./encoders.js";
-import { makeLZ4Encoder } from "./lz4Encoder.js";
+import {
+	type IProvideRuntimeAttributor,
+	type IRuntimeAttributor,
+} from "./attributorContracts.js";
+import { RuntimeAttributor } from "./runtimeAttributor.js";
 
 /**
  * Data store channel for the runtime attributor. This channel is responsible for storing and managing the
  */
 export class RuntimeAttributorDataStoreChannel
 	extends TypedEventEmitter<IFluidDataStoreRuntimeEvents>
-	implements IFluidDataStoreChannel, IRuntimeAttributor
+	implements IFluidDataStoreChannel, IProvideRuntimeAttributor
 {
 	public constructor(
 		public readonly dataStoreContext: IFluidDataStoreContext,
 		existing: boolean,
 	) {
 		super();
+		this.runtimeAttributor = new RuntimeAttributor();
 		this.mc = createChildMonitoringContext({
 			logger: dataStoreContext.baseLogger,
 			namespace: "Attributor",
@@ -75,7 +74,7 @@ export class RuntimeAttributorDataStoreChannel
 			this.deferredAttached.resolve();
 		}
 		this.entryPoint = new FluidObjectHandle<FluidObject>(
-			this,
+			this.runtimeAttributor,
 			"",
 			dataStoreContext.IFluidHandleContext,
 		);
@@ -86,47 +85,8 @@ export class RuntimeAttributorDataStoreChannel
 	}
 
 	public get IRuntimeAttributor(): IRuntimeAttributor {
-		return this;
+		return this.runtimeAttributor;
 	}
-
-	public get(key: AttributionKey): AttributionInfo {
-		assert(
-			this.opAttributor !== undefined,
-			0x509 /* RuntimeAttributor must be initialized before getAttributionInfo can be called */,
-		);
-
-		if (key.type === "detached") {
-			throw new Error("Attribution of detached keys is not yet supported.");
-		}
-
-		if (key.type === "local") {
-			// Note: we can *almost* orchestrate this correctly with internal-only changes by looking up the current
-			// client id in the audience. However, for read->write client transition, the container might have not yet
-			// received a client id. This is left as a TODO as it might be more easily solved once the detached case
-			// is settled (e.g. if it's reasonable for the host to know the current user information at container
-			// creation time, we could just use that here as well).
-			throw new Error("Attribution of local keys is not yet supported.");
-		}
-
-		return this.opAttributor.getAttributionInfo(key.seq);
-	}
-
-	public has(key: AttributionKey): boolean {
-		if (key.type === "detached") {
-			return false;
-		}
-
-		if (key.type === "local") {
-			return false;
-		}
-
-		return this.opAttributor?.tryGetAttributionInfo(key.seq) !== undefined;
-	}
-
-	private encoder: Encoder<IAttributor, string> = {
-		encode: unreachableCase,
-		decode: unreachableCase,
-	};
 
 	private _disposed = false;
 	public get disposed(): boolean {
@@ -137,7 +97,7 @@ export class RuntimeAttributorDataStoreChannel
 		this._disposed = true;
 	}
 
-	private opAttributor: IAttributor | undefined;
+	private readonly runtimeAttributor: RuntimeAttributor;
 	public isEnabled = true;
 	public _attachState: AttachState;
 	public visibilityState: VisibilityState;
@@ -153,26 +113,12 @@ export class RuntimeAttributorDataStoreChannel
 		baseSnapshotForAttributorTree: ISnapshotTree | undefined,
 		readBlob: (id: string) => Promise<ArrayBufferLike>,
 	): Promise<void> {
-		this.encoder = chain(
-			new AttributorSerializer(
-				(entries) => new OpStreamAttributor(deltaManager, quorum, entries),
-				deltaEncoder,
-			),
-			makeLZ4Encoder(),
+		await this.runtimeAttributor.initialize(
+			deltaManager,
+			quorum,
+			baseSnapshotForAttributorTree,
+			readBlob,
 		);
-
-		if (baseSnapshotForAttributorTree === undefined) {
-			this.opAttributor = new OpStreamAttributor(deltaManager, quorum);
-		} else {
-			const id = baseSnapshotForAttributorTree.blobs[opBlobName];
-			assert(
-				id !== undefined,
-				0x50a /* Attributor tree should have op attributor summary blob. */,
-			);
-			const blobContents = await readBlob(id);
-			const attributorSnapshot = bufferToString(blobContents, "utf8");
-			this.opAttributor = this.encoder.decode(attributorSnapshot);
-		}
 	}
 
 	public attachGraph(): void {
@@ -199,13 +145,7 @@ export class RuntimeAttributorDataStoreChannel
 	 * {@inheritdoc IFluidDataStoreChannel.getAttachSummary}
 	 */
 	public getAttachSummary(telemetryContext?: ITelemetryContext): ISummaryTreeWithStats {
-		assert(
-			this.opAttributor !== undefined,
-			0x50b /* RuntimeAttributor should be initialized before summarization */,
-		);
-		const builder = new SummaryTreeBuilder();
-		builder.addBlob(opBlobName, this.encoder.encode(this.opAttributor));
-		return builder.getSummaryTree();
+		return this.runtimeAttributor.summarizeOpAttributor();
 	}
 
 	/**
@@ -241,13 +181,7 @@ export class RuntimeAttributorDataStoreChannel
 		trackState?: boolean,
 		telemetryContext?: ITelemetryContext,
 	): Promise<ISummaryTreeWithStats> {
-		assert(
-			this.opAttributor !== undefined,
-			"RuntimeAttributor should be initialized before summarization",
-		);
-		const builder = new SummaryTreeBuilder();
-		builder.addBlob(opBlobName, this.encoder.encode(this.opAttributor));
-		return builder.getSummaryTree();
+		return this.runtimeAttributor.summarizeOpAttributor();
 	}
 
 	/**
@@ -320,8 +254,6 @@ export class RuntimeAttributorDataStoreChannel
 	public setAttachState(attachState: AttachState.Attaching | AttachState.Attached): void {
 		switch (attachState) {
 			case AttachState.Attaching: {
-				// this.attachGraph();
-
 				this._attachState = AttachState.Attaching;
 
 				assert(
